更新：后端/系统模块、闭环、GPS、轨迹与架构文档
- 架构规则与后端阻塞设计文档
- 增量优化器、ISAM2 因子、任务派发与 HBA 相关
- 闭环检测/ICP、子图、GPS 处理与轨迹日志
- 新增后端模块化分析与 finish_mapping 根因文档

Made-with: Cursor

+++ automap_pro/src/loop_closure/icp_refiner.cpp
@@ -323,7 +323,7 @@ Mat66d IcpRefiner::computeCorrespondenceCovariance(
     
     // 计算对应误差的协方差
     for (const auto& [src_idx, tgt_idx] : correspondences) {
-        if (src_idx >= src->size() || tgt_idx >= tgt->size()) continue;
+        if (static_cast<std::size_t>(src_idx) >= src->size() || static_cast<std::size_t>(tgt_idx) >= tgt->size()) continue;
         
         const auto& src_pt = src->points[src_idx];
         const auto& tgt_pt = tgt->points[tgt_idx];
@@ -453,7 +453,7 @@ void IcpRefiner::filterCorrespondences(
     std::vector<std::pair<int, int>> filtered;
     
     for (const auto& [src_idx, tgt_idx] : correspondences) {
-        if (src_idx >= src->size() || tgt_idx >= tgt->size()) continue;
+        if (static_cast<std::size_t>(src_idx) >= src->size() || static_cast<std::size_t>(tgt_idx) >= tgt->size()) continue;
         
         const auto& src_pt = src->points[src_idx];
         const auto& tgt_pt = tgt->points[tgt_idx];
